requirements.txt
fastapi
uvicorn
supabase
pymupdf
anthropic
sendgrid
apscheduler
python-dotenv
httpx
jinja2
python-multipart
requests